% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{agingTest}
\alias{agingTest}
\title{Calculates the number of times an indicator is repeated on a feed throughout}
\usage{
agingTest(group, start.date, end.date, type = "raw", split.ti = TRUE,
  select.sources = NULL, .progress = TRUE)
}
\arguments{
\item{group}{the name of the dataset group. Must exist on the 'type' category}

\item{start.date}{the beginning date for the test}

\item{end.date}{the end date for the test}

\item{type}{The test can take into consideration the FQDN sources as
the original entities ("raw"), or as the extracted IPv4 fields from
the enriched entities ("enriched")}

\item{split.ti}{if TRUE, creates a popoulation for each source and returns a list
with the sources as IDs. Otherwise, returns a list with a single
element with the group as the ID.}

\item{select.sources}{a chararacter vector of the sources on the dataset you want
to be a part of the test, or NULL for all of them. Only
applicable if split.ti = TRUE.}

\item{.progress}{show a progress bar? (default: \code{TRUE} - show a progress bar)}
}
\value{
"agingtest" object
}
\description{
Calculates the number of times an indicator is repeated on a feed throughout
the days from 'start.date' to 'end.date' Use 'group' and 'type' to select the
dataset. 'split.ti' and 'select.sources' control the output.\cr
\cr
Defaults to showing a progress bar. Set \code{.progress} to \code{FALSE} when knitting
}

