% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{extractPopulationFromTI}
\alias{extractPopulationFromTI}
\title{Returns multiple population data.tables calculated using the sources on the
"enriched" TI dataset on 'date'.}
\usage{
extractPopulationFromTI(group, pop.id, date, split.ti = TRUE,
  select.sources = NULL)
}
\arguments{
\item{group}{the name of the dataset group. Must exist on "enriched" category}

\item{pop.id}{the key of the population dataset to generate. Can be multiple keys}

\item{date}{the date you want to use}

\item{split.ti}{if TRUE, creates a popoulation for each source and returns a list
with the sources as IDs. Otherwise, returns a list with a single
element with the group as the ID.}

\item{select.sources}{a chararacter vector of the sources on the dataset you want}
}
\value{
'list' of population data.tables
}
\description{
Use 'group' to select the dataset, and 'pop.id' for the population key
aggregation metric. 'split.ti' and 'select.sources' control the output
}

