% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-data.R
\name{getAvailableDates}
\alias{getAvailableDates}
\title{Queries the data repository for the specific 'category' and 'group' of TI
or population information and return the dates we have available for that
group.}
\usage{
getAvailableDates(category, group)
}
\arguments{
\item{category}{currently one of "raw", "enriched", "population"}

\item{group}{the intel group you want the path for}
}
\value{
'character' vector
}
\description{
Queries the data repository for the specific 'category' and 'group' of TI
or population information and return the dates we have available for that
group.
}

