% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-data.R
\name{getDirPath}
\alias{getDirPath}
\title{Builds the path for the intel analysts sources and returns it.}
\usage{
getDirPath(category, group)
}
\arguments{
\item{category}{currently one of "raw", "enriched", "population"}

\item{group}{the intel group you want the path for}
}
\value{
'character'
}
\description{
Builds the path for the intel analysts sources and returns it.
}

