% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-data.R
\name{loadData}
\alias{loadData}
\title{Lower level function helper for loadTI and loadPopulation. Should not be used
directly}
\usage{
loadData(category, group, date = NULL, valid.fields = character(0))
}
\arguments{
\item{category}{currently one of "raw", "enriched" or "population"}

\item{group}{the intel group you want the path for}

\item{datea}{date string in the format "YYYYMMDD", or NULL if you want the
date the 'database' has available}

\item{valid.fields:}{the fields we should look for in the dataset to validate that
it is of the correct type}
}
\value{
a 'data.table'
}
\description{
Lower level function helper for loadTI and loadPopulation. Should not be used
directly
}

