% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-data.R
\name{loadPopulation}
\alias{loadPopulation}
\title{Fetches the data contained in the Population dataset contained in
'pop.group', with the 'pop.id' as the key on the 'date' specified.}
\usage{
loadPopulation(pop.group, pop.id, date = NULL)
}
\arguments{
\item{pop.group}{the intel group you want the path for}

\item{pop.id}{the key of the population dataset. If not present in the dataset
the function will return NULL.}

\item{date}{a date string in the format "YYYYMMDD", or NULL if you want the
date the 'database' has available}
}
\value{
'data.table' with 'pop.id' and 'totalIPs'
}
\description{
Fetches the data contained in the Population dataset contained in
'pop.group', with the 'pop.id' as the key on the 'date' specified.
}

