% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-data.R
\name{loadTI}
\alias{loadTI}
\title{Fetches the data contained in the Threat Intelligence dataset contained in
'category' and 'group' on the 'date' specified.}
\usage{
loadTI(category, group, date = NULL)
}
\arguments{
\item{category}{currently one of "raw" or "enriched"}

\item{group}{the intel group you want the path for}

\item{date}{a date string in the format "YYYYMMDD", or NULL if you want the
date the 'database' has available}
}
\value{
data.table' with at least '.tiq.data.defaultTIFields'
}
\description{
Fetches the data contained in the Threat Intelligence dataset contained in
'category' and 'group' on the 'date' specified.
}

