% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{noveltyTest}
\alias{noveltyTest}
\title{The novelty test is about how many indicators are being added and removed
on each passing day on a specific TI feed.}
\usage{
noveltyTest(group, start.date, end.date, select.sources = NULL,
  .progress = TRUE)
}
\arguments{
\item{group}{the name of the dataset group. This test works exclusively on "raw"}

\item{start.date}{the beginning date for the test}

\item{end.date}{the end date for the test}

\item{select.sources}{a chararacter vector of the sources on the dataset you want
to be a part of the test, or NULL for all of them}

\item{.progress}{show a progress bar? (default: \code{TRUE} - show a progress bar)}
}
\value{
a 'list'
}
\description{
The comparisons are done on each 'source' over the multiple days they are
available. \cr
\cr
Calculates the novelty test results for a specific 'group' of TI datasets from
the 'start.date' to the 'end.date'. You need at least 2 days to have a comparison.
You can alternatively select what sources will be a part of the test by
providing a character vector to 'select.sources'\cr
\cr
Defaults to showing a progress bar. Set \code{.progress} to \code{FALSE} when knitting
}

