% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{overlapTest}
\alias{overlapTest}
\title{The overlap test is about how many indicators are present and repeted on
multiple TI feeds.}
\usage{
overlapTest(group, date, type = "raw", select.sources = NULL)
}
\arguments{
\item{group}{the name of the dataset group. Must exist on the selected 'type'}

\item{date}{the date you want the test to run in}

\item{type}{The overlap test can take into consideration the FQDN sources as
the original entities ("raw"), or as the extracted IPv4 fields from
the enriched entities ("enriched")}

\item{select.sources}{a chararacter vector of the sources on the dataset you want
to be a part of the test, or NULL for all of them}
}
\value{
a numeric matrix with the overlap ratios
}
\description{
The comparisons are done on between each "source" selected
on the same group on a specific day
}

