% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{plotAgingTest}
\alias{plotAgingTest}
\title{Plots an Aging Test histogram and density plot}
\usage{
plotAgingTest(aging.data, title = NULL, plot.sources = NULL)
}
\arguments{
\item{aging.data}{the aging data object to be plotted. More specifically the output of
tiq.test.agingTest()}

\item{title:}{a title for your plot. NULL leaves it blank}

\item{plot.sources:}{a chararacter vector of the sources on the novelty test you want
to be a part of the plot, or NULL for all of them}
}
\description{
Plots an Aging Test histogram and density plot
}

