% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{plotNoveltyTest}
\alias{plotNoveltyTest}
\title{Plots the results of the novelty test in a (mostly) clear graph for comparisons}
\usage{
plotNoveltyTest(novelty, sources = NULL)
}
\arguments{
\item{novelty}{the output of the 'tiq.test.noveltyTest' function}

\item{plot.sources}{a chararacter vector of the sources on the novelty test you want
to be a part of the plot, or NULL for all of them}
}
\description{
Plots the results of the novelty test in a (mostly) clear graph for comparisons
}

