% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{plotOverlapTest}
\alias{plotOverlapTest}
\title{Plots the results of the overlap test in a (mostly) clear heatmap for comparisons}
\usage{
plotOverlapTest(overlap, title = "Overlap Test Plot", plot.sources = NULL)
}
\arguments{
\item{overlap}{the output of the 'tiq.test.OverlapTest' function}

\item{title}{a title for your plot}

\item{plot.sources}{a chararacter vector of the sources on the novelty test you want
to be a part of the plot, or NULL for all of them}
}
\value{
ggplot2 object (plots when printed)
}
\description{
Plots the results of the overlap test in a (mostly) clear heatmap for comparisons
}

