% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{plotPopulationBars}
\alias{plotPopulationBars}
\title{Plots a population bar chart for simple comparisons}
\usage{
plotPopulationBars(pop.data, pop.id, table.size = 10, title = "",
  plot.sources = NULL)
}
\arguments{
\item{pop.data}{the population data table to be plotted}

\item{pop.id}{the id of the population dataset}

\item{table.size}{the number of bars on the graph. Try not to go too crazy!}

\item{title}{a title for your plot}

\item{plot.sources}{a chararacter vector of the sources on the novelty test you want
to be a part of the plot, or NULL for all of them}
}
\description{
Plots a population bar chart for simple comparisons
}

