% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-test.R
\name{populationInference}
\alias{populationInference}
\title{Runs an inference-based comparison between the reference population ('ref.pop')
and the test population ('test.pop'), based on the 'pop.id' key.}
\usage{
populationInference(ref.pop, test.pop, pop.id, exact = TRUE, top = 25)
}
\arguments{
\item{ref.pop}{a population dataset (not on a list) from the reference population}

\item{test.pop}{a population dataset (not on a list) to compare}

\item{pop.id}{the id of the population dataset. Supports multiple ids, indexes
the output by the first one}

\item{exact}{When this is TRUE (the default), the function will execute an
exact binomial test with the proportion on ref.pop. This should
only be the case when ref.pop is the ACTUAL population (e.g. from
MaxMind GeoIP database). Otherwise, set this to FALSE for a
chi-squared test to compare the different proportions}

\item{top}{the X top members of the test.pop we want to test. Set to -1 for all.}
}
\value{
a data.table
}
\description{
You should use the exact option when the ref.pop was obtained from a
"population" dataset, and the other one when it is extracted from another
TI feed. Also, the "not-exact" test loses a lot of precision if the numbers
are too low, so stick with the Top X members od the test.pop population ideally
}

