% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tiq-package.R
\docType{package}
\name{tiq}
\alias{tiq}
\alias{tiq-package}
\title{Functions for the Threat Intelligence IQ Tests. These should be used for the
different evaluations to be run on the TI feeds \cr
\cr
For more information, please refer to the README at
\url{https://github.com/mlsecproject/tiq-test/}}
\description{
Functions for the Threat Intelligence IQ Tests. These should be used for the
different evaluations to be run on the TI feeds \cr
\cr
For more information, please refer to the README at
\url{https://github.com/mlsecproject/tiq-test/}
}
\author{
Alex Pinto (@alexcpsec)
}

